refactor(agents): split tool search internals
