+++ src/agents/tool-search-runtime.ts
@@ -1,0 +1,429 @@
+import { isRecord } from "@openclaw/normalization-core/record-coerce";
+import {
+  normalizeStringEntries,
+  uniqueStrings,
+} from "@openclaw/normalization-core/string-normalization";
+import { getPluginToolMeta } from "../plugins/tools.js";
+import { isPreExecutionBlockedToolResult } from "./agent-tools.before-tool-call.js";
+import { getChannelAgentToolMeta } from "./channel-tool-metadata.js";
+import type { AgentToolResult } from "./runtime/index.js";
+import { isAgentToolReplaySafe } from "./tool-replay-safety.js";
+import {
+  compactToolSearchCatalogEntry,
+  resolveCatalog,
+  visibleCatalogEntries,
+} from "./tool-search-catalog.js";
+import { snapshotToolSearchTargetTranscriptResult } from "./tool-search-transcript.js";
+import type {
+  CatalogSource,
+  CatalogVisibilityOptions,
+  ToolSearchCallOptions,
+  ToolSearchCatalogEntry,
+  ToolSearchCatalogSession,
+  ToolSearchCatalogToolExecutor,
+  ToolSearchConfig,
+  ToolSearchToolContext,
+  UnknownToolErrorOptions,
+  UnknownToolRecoverySurface,
+} from "./tool-search-types.js";
+import { asToolParamsRecord, ToolInputError } from "./tools/common.js";
+
+function describeEntry(entry: ToolSearchCatalogEntry) {
+  return {
+    ...compactToolSearchCatalogEntry(entry),
+    parameters: entry.parameters ?? {},
+    ...(entry.outputSchema ? { outputSchema: entry.outputSchema } : {}),
+  };
+}
+
+function tokenize(input: string): string[] {
+  return normalizeStringEntries(input.toLowerCase().split(/[^a-z0-9_./:-]+/u));
+}
+
+function scoreEntry(entry: ToolSearchCatalogEntry, terms: string[]): number {
+  if (terms.length === 0) {
+    return 1;
+  }
+  const name = entry.name.toLowerCase();
+  const id = entry.id.toLowerCase();
+  const label = (entry.label ?? "").toLowerCase();
+  const description = entry.description.toLowerCase();
+  let score = 0;
+  for (const term of terms) {
+    if (name === term || id === term) {
+      score += 20;
+    }
+    if (name.includes(term)) {
+      score += 8;
+    }
+    if (id.includes(term)) {
+      score += 6;
+    }
+    if (label.includes(term)) {
+      score += 4;
+    }
+    if (description.includes(term)) {
+      score += 2;
+    }
+  }
+  return score;
+}
+
+function tokenizeLookupValue(input: string): Set<string> {
+  return new Set(normalizeStringEntries(input.toLowerCase().split(/[^a-z0-9]+/u)));
+}
+
+function scoreUnknownToolSuggestion(needle: string, entry: ToolSearchCatalogEntry): number {
+  const normalizedNeedle = needle.toLowerCase();
+  const name = entry.name.toLowerCase();
+  const id = entry.id.toLowerCase();
+  const label = (entry.label ?? "").toLowerCase();
+  const description = entry.description.toLowerCase();
+  const needleTokens = tokenizeLookupValue(needle);
+  const entryTokens = tokenizeLookupValue(
+    `${entry.name} ${entry.id} ${entry.label ?? ""} ${entry.description}`,
+  );
+  let score = 0;
+  if ((name && normalizedNeedle.includes(name)) || id.includes(normalizedNeedle)) {
+    score += 40;
+  }
+  if (name && needleTokens.has(name)) {
+    score += 40;
+  }
+  for (const token of needleTokens) {
+    if (entryTokens.has(token)) {
+      score += 12;
+    }
+  }
+  if (label.includes(normalizedNeedle) || description.includes(normalizedNeedle)) {
+    score += 8;
+  }
+  return score;
+}
+
+function formatUnknownToolIdError(
+  needle: string,
+  entries: readonly ToolSearchCatalogEntry[],
+  options: UnknownToolErrorOptions = {},
+): string {
+  const nameCounts = new Map<string, number>();
+  for (const entry of entries) {
+    nameCounts.set(entry.name, (nameCounts.get(entry.name) ?? 0) + 1);
+  }
+  const suggestions = uniqueStrings(
+    entries
+      .map((entry) => ({
+        value: options.exactIdOnly || (nameCounts.get(entry.name) ?? 0) > 1 ? entry.id : entry.name,
+        score: scoreUnknownToolSuggestion(needle, entry),
+      }))
+      .filter((candidate) => candidate.score > 0)
+      .toSorted((a, b) => b.score - a.score || a.value.localeCompare(b.value))
+      .map((candidate) => candidate.value),
+  ).slice(0, 3);
+  const recoveryText =
+    options.recoverySurface === "code-mode"
+      ? "Use openclaw.tools.search to find a tool, openclaw.tools.describe to inspect it, then openclaw.tools.call with the exact id or name."
+      : options.recoverySurface === "tools"
+        ? "Use tools.search to find a tool, tools.describe to inspect it, then tools.call with the exact id or name."
+        : "Use tool_search to find a tool, tool_describe to inspect it, then tool_call with the exact id or name.";
+  if (suggestions.length === 0) {
+    return `Unknown tool id: ${needle}. ${recoveryText}`;
+  }
+  return `Unknown tool id: ${needle}. Did you mean: ${suggestions.join(", ")}? ${recoveryText}`;
+}
+
+function findEntry(
+  catalog: ToolSearchCatalogSession,
+  id: string,
+  options?: CatalogVisibilityOptions,
+  errorOptions?: UnknownToolErrorOptions,
+): ToolSearchCatalogEntry {
+  const needle = id.trim();
+  const entries = visibleCatalogEntries(catalog, options);
+  const exactIdEntry = entries.find((candidate) => candidate.id === needle);
+  if (exactIdEntry) {
+    return exactIdEntry;
+  }
+  const namedEntries = entries.filter((candidate) => candidate.name === needle);
+  if (namedEntries.length > 1) {
+    throw new ToolInputError(`Ambiguous tool name: ${needle}; use an exact tool id.`);
+  }
+  const namedEntry = namedEntries[0];
+  if (!namedEntry) {
+    throw new ToolInputError(formatUnknownToolIdError(needle, entries, errorOptions));
+  }
+  return namedEntry;
+}
+
+function findEntryByExactId(
+  catalog: ToolSearchCatalogSession,
+  id: string,
+  errorOptions: UnknownToolErrorOptions = {},
+): ToolSearchCatalogEntry {
+  const needle = id.trim();
+  const entry = catalog.entries.find((candidate) => candidate.id === needle);
+  if (!entry) {
+    throw new ToolInputError(
+      formatUnknownToolIdError(needle, catalog.entries, { ...errorOptions, exactIdOnly: true }),
+    );
+  }
+  return entry;
+}
+
+export function readToolSearchId(args: unknown): string {
+  const params = asToolParamsRecord(args);
+  const value = params.id ?? params.toolId ?? params.name;
+  if (typeof value !== "string" || !value.trim()) {
+    throw new ToolInputError("id must be a non-empty string.");
+  }
+  return value.trim();
+}
+
+function readToolSearchLimit(value: unknown, config: ToolSearchConfig): number {
+  if (value === undefined) {
+    return config.searchDefaultLimit;
+  }
+  if (typeof value !== "number" || !Number.isInteger(value) || value < 1) {
+    throw new ToolInputError("limit must be a positive integer.");
+  }
+  return Math.min(value, config.maxSearchLimit);
+}
+
+export function readToolSearchArgs(
+  args: unknown,
+  config: ToolSearchConfig,
+): { query: string; limit: number } {
+  const params = asToolParamsRecord(args);
+  const query = params.query;
+  if (typeof query !== "string") {
+    throw new ToolInputError("query must be a string.");
+  }
+  const options = isRecord(params.options) ? params.options : undefined;
+  return {
+    query,
+    limit: readToolSearchLimit(params.limit ?? options?.limit, config),
+  };
+}
+
+export function readToolSearchCallArgs(args: unknown): { id: string; input: unknown } {
+  const params = asToolParamsRecord(args);
+  return {
+    id: readToolSearchId(params),
+    input: params.args ?? params.input ?? {},
+  };
+}
+
+function getTelemetry(catalog: ToolSearchCatalogSession) {
+  const sources: Record<CatalogSource, number> = { openclaw: 0, mcp: 0, client: 0 };
+  for (const entry of catalog.entries) {
+    sources[entry.source] += 1;
+  }
+  return {
+    catalogSize: catalog.entries.length,
+    sources,
+    searchCount: catalog.searchCount,
+    describeCount: catalog.describeCount,
+    callCount: catalog.callCount,
+  };
+}
+
+let schemaValidatorModulePromise:
+  | Promise<typeof import("../plugins/schema-validator.js")>
+  | undefined;
+
+async function validateCatalogOutputValue(
+  entry: ToolSearchCatalogEntry,
+  value: unknown,
+): Promise<
+  ReturnType<typeof import("../plugins/schema-validator.js").validateJsonSchemaValue> | undefined
+> {
+  if (!entry.outputSchema) {
+    return undefined;
+  }
+  try {
+    schemaValidatorModulePromise ??= import("../plugins/schema-validator.js");
+    const { validateJsonSchemaValue } = await schemaValidatorModulePromise;
+    return validateJsonSchemaValue({
+      schema: entry.outputSchema as never,
+      cacheKey: `tool-output:${entry.id}`,
+      value,
+    });
+  } catch (error) {
+    throw new Error(`Tool "${entry.id}" has an invalid outputSchema.`, { cause: error });
+  }
+}
+
+async function assertCatalogOutputSchemaIsValid(entry: ToolSearchCatalogEntry): Promise<void> {
+  // Compile before execution so a bad contract cannot follow a successful side effect.
+  await validateCatalogOutputValue(entry, undefined);
+}
+
+async function assertCatalogOutputMatchesSchema(
+  entry: ToolSearchCatalogEntry,
+  result: AgentToolResult<unknown>,
+): Promise<void> {
+  if (!entry.outputSchema) {
+    return;
+  }
+  if (isPreExecutionBlockedToolResult(result)) {
+    const details = unwrapToolResultValue(result);
+    const reason =
+      isRecord(details) && typeof details.reason === "string" && details.reason.trim()
+        ? details.reason
+        : "Tool call blocked by policy";
+    throw new Error(`Tool "${entry.id}" was blocked before execution: ${reason}`);
+  }
+  const validation = await validateCatalogOutputValue(entry, unwrapToolResultValue(result));
+  if (!validation || validation.ok) {
+    return;
+  }
+  throw new Error(
+    `Tool "${entry.id}" returned details that do not match its declared outputSchema.`,
+  );
+}
+
+function sanitizeToolCallIdPart(value: string): string {
+  const trimmed = value.trim();
+  const safe = trimmed.replace(/[^A-Za-z0-9_.:-]+/g, "_").slice(0, 120);
+  return safe || "call";
+}
+
+export class ToolSearchRuntime {
+  private callSequence = 0;
+
+  constructor(
+    private readonly ctx: ToolSearchToolContext,
+    private readonly config: ToolSearchConfig,
+  ) {}
+
+  search = async (query: string, options?: { limit?: number } & CatalogVisibilityOptions) => {
+    const catalog = resolveCatalog(this.ctx);
+    catalog.searchCount += 1;
+    const limit = readToolSearchLimit(options?.limit, this.config);
+    const terms = tokenize(query);
+    return visibleCatalogEntries(catalog, options)
+      .map((entry) => ({ entry, score: scoreEntry(entry, terms) }))
+      .filter((hit) => hit.score > 0)
+      .toSorted((a, b) => b.score - a.score || a.entry.id.localeCompare(b.entry.id))
+      .slice(0, limit)
+      .map((hit) => compactToolSearchCatalogEntry(hit.entry));
+  };
+
+  all = (options?: CatalogVisibilityOptions) =>
+    visibleCatalogEntries(resolveCatalog(this.ctx), options).map((entry) =>
+      compactToolSearchCatalogEntry(entry),
+    );
+
+  namespaceEntries = () =>
+    resolveCatalog(this.ctx).entries.map((entry) =>
+      Object.assign(compactToolSearchCatalogEntry(entry), {
+        parameters: entry.parameters ?? {},
+      }),
+    );
+
+  describe = async (id: string, options?: CatalogVisibilityOptions & UnknownToolErrorOptions) => {
+    const catalog = resolveCatalog(this.ctx);
+    catalog.describeCount += 1;
+    return describeEntry(findEntry(catalog, id, options, options));
+  };
+
+  call = async (id: string, input?: unknown, options?: ToolSearchCallOptions) => {
+    const catalog = resolveCatalog(this.ctx);
+    return await this.callEntry(catalog, findEntry(catalog, id, options, options), input, options);
+  };
+
+  callExactId = async (
+    id: string,
+    input?: unknown,
+    options?: {
+      parentToolCallId?: string;
+      signal?: AbortSignal;
+      onUpdate?: ToolSearchCallOptions["onUpdate"];
+      recoverySurface?: UnknownToolRecoverySurface;
+    },
+  ) => {
+    const catalog = resolveCatalog(this.ctx);
+    return await this.callEntry(catalog, findEntryByExactId(catalog, id, options), input, options);
+  };
+
+  callValue = async (id: string, input?: unknown, options?: ToolSearchCallOptions) =>
+    unwrapToolResultValue((await this.call(id, input, options)).result);
+
+  isReplaySafeExactId = (id: string): boolean => {
+    let entry: ToolSearchCatalogEntry;
+    try {
+      entry = findEntryByExactId(resolveCatalog(this.ctx), id);
+    } catch {
+      return false;
+    }
+    if (entry.source !== "openclaw") {
+      return false;
+    }
+    const pluginMeta = getPluginToolMeta(entry.tool as Parameters<typeof getPluginToolMeta>[0]);
+    if (pluginMeta) {
+      return pluginMeta.mcp ? false : pluginMeta.replaySafe === true;
+    }
+    if (getChannelAgentToolMeta(entry.tool as never)) {
+      return false;
+    }
+    return isAgentToolReplaySafe(entry.tool);
+  };
+
+  private readonly callEntry = async (
+    catalog: ToolSearchCatalogSession,
+    entry: ToolSearchCatalogEntry,
+    input?: unknown,
+    options?: {
+      parentToolCallId?: string;
+      signal?: AbortSignal;
+      onUpdate?: ToolSearchCallOptions["onUpdate"];
+    },
+  ) => {
+    catalog.callCount += 1;
+    await assertCatalogOutputSchemaIsValid(entry);
+    const parentId = sanitizeToolCallIdPart(options?.parentToolCallId ?? "direct");
+    const toolCallId = `tool_search_code:${parentId}:${entry.name}:${++this.callSequence}`;
+    const executeTool =
+      this.ctx.executeTool ??
+      (async (params: Parameters<ToolSearchCatalogToolExecutor>[0]) => {
+        const result = await params.tool.execute(
+          params.toolCallId,
+          params.input,
+          params.signal,
+          params.onUpdate,
+          undefined as never,
+        );
+        return await params.acceptResultBeforeProjection(result);
+      });
+    const acceptResultBeforeProjection = async (candidate: AgentToolResult<unknown>) => {
+      if (isPreExecutionBlockedToolResult(candidate)) {
+        await assertCatalogOutputMatchesSchema(entry, candidate);
+      }
+      const snapshot = snapshotToolSearchTargetTranscriptResult(candidate);
+      await assertCatalogOutputMatchesSchema(entry, snapshot);
+      return snapshot;
+    };
+    const result = await executeTool({
+      tool: entry.tool,
+      toolName: entry.name,
+      source: entry.source,
+      sourceName: entry.sourceName,
+      toolCallId,
+      parentToolCallId: options?.parentToolCallId,
+      input: input ?? {},
+      signal: options?.signal ?? this.ctx.abortSignal,
+      onUpdate: options?.onUpdate,
+      acceptResultBeforeProjection,
+    });
+    const acceptedResult = await acceptResultBeforeProjection(result);
+    return { tool: compactToolSearchCatalogEntry(entry), result: acceptedResult };
+  };
+
+  telemetry() {
+    return getTelemetry(resolveCatalog(this.ctx));
+  }
+}
+
+function unwrapToolResultValue(result: AgentToolResult<unknown>): unknown {
+  return isRecord(result) && "details" in result ? result.details : result;
+}
